requirements.txt
transformers
datasets
accelerate
google-generativeai
bs4